API auth directly in SocialSecure. Now a POST value. Resource stubs.
How to login:
POST request on /api/v1/auth?provider=userpass
with POST parameters:
username=email@email.com
password=password

Should be pretty much final for the how to connect since we now receive a cookie. The only downside is that it needed a scala controller to have all the specific stuff (It's basicly the ProviderController in SS without the redirections).

Added some Resource controllers.

Signed-off-by: William Lebel <lebelwilliam@gmail.com>